pylibs/requirements.txt
nose
-e git://github.com/nvie/nose-machineout.git#egg=nose_machineout
pylint
pyflakes
rope
path.py
